chore: Enable unconvert linter
14 redundant type conversions removed:
- 9 string(data)/string(result.data) where Data is a type alias to string
- 4 int64(p.Experience)/int64(experience) where Experience is int64
- 1 int64(achievement.After.Player.Experience)

Lint time: unchanged (~0.5s).

+++ internal/adapters/playerrepository/repository.go
@@ -151,7 +151,7 @@ func dbStatToPlayerPIT(dbStat dbStat) (*domain.PlayerPIT, error) {
 		LastLogin:   nil,
 		LastLogout:  nil,
 
-		Experience: int64(experience),
+		Experience: experience,
 		Solo:       *gamemodeStatsPITFromDataStorage(&playerData.Solo),
 		Doubles:    *gamemodeStatsPITFromDataStorage(&playerData.Doubles),
 		Threes:     *gamemodeStatsPITFromDataStorage(&playerData.Threes),
